Added submodel handling
Submodels are added with the addSubmodel() method. Whatever is passed as the name is the column to be used as the foreign key. The values in this column are substituted during select with an array of any rows in the submodels having that value in the column identified with the submodel's primaryKey property.