Allow Transactions to be added as "queries"
A Transaction can now be invoked from inside another Transaction as if it is a normal procedure (like a Query, PreparedStatement, or StatementSet), and can be referenced from a function defined in the outer Transaction's addFunction method. When the outer Transaction is run, the inner Transaction is run at its designated time, without committing. Savepoints are created at the beginning of every Transaction, so that the inner Transaction can be rolled back as an atomic unit if needed.