Timestamp file creation removed
The timestamp file was meant to keep track of last-modified times per table per database for caching purposes. PrivateTmp=true interferes with this in a way that is difficult to work around without modifying system configuration files. This will be revisited later, but is not needed for this phase of development.