Coerce all numeric values in output to strings
json_encode() behavior seems to have changed in PHP 8.1 (this isn't in the release notes?); prior to this version, all values were output as strings, regardless of whether they were numeric. This new behavior will break any front-end code that depends on strict comparisons (=== or !===) where a numeric value is being compared to a string. To keep this from being an issue, all numbers will be converted to strings prior to output.