Added passthru property for select Diffs
Passthru property allows criteria to be applied directly to the called StatementSet rather than the resulting Model. The data in the resulting Model will be limited by these criteria regardless of what operations are performed on it after the fact. A passthru call is made by adding a passthru property to the select object passed from the client. (the value doesn't matter as long as it's not falsy)