Added non-standard EKIL / EKILR operators
These are inverted LIKE / RLIKE operators, respectively. When these are used, they are converted to their non-inverted counterparts, but the comparison is reversed (putting the value on the left and the column on the right). This is useful when comparing a string against a pattern stored in the given column.